simplydigitals-aiconnoisseur-api/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[project]
name = "mlapi"
version = "1.0.0"
description = "Production ML Analytics REST API Platform"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "ML Platform Team" }]

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "python-multipart>=0.0.9",

    # ORM & Database
    "sqlalchemy>=2.0.30",
    "alembic>=1.13.1",
    "psycopg2-binary>=2.9.9",
    "aiosqlite>=0.20.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "slowapi>=0.1.9",

    # Validation
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "email-validator>=2.1.1",

    # ML / Data Science
    "scikit-learn>=1.4.2",
    "pandas>=2.2.2",
    "numpy>=1.26.4",
    "joblib>=1.4.0",

    # Observability
    "structlog>=24.1.0",
    "prometheus-fastapi-instrumentator>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=3.7.0",
    "ruff>=0.4.4",
    "black>=24.4.2",
    "mypy>=1.10.0",
    "bandit[toml]>=1.7.8",
    "safety>=3.2.0",
    "types-passlib>=1.7.7.20240327",
    "types-python-jose>=3.3.4.20240106",
]
test = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "behave>=1.2.6",
    "factory-boy>=3.3.0",
    "faker>=25.0.0",
]

# ─────────────────────────────────────────────────────────────
# Tool Configuration
# ─────────────────────────────────────────────────────────────

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # flake8-bandit (security)
    "N",   # pep8-naming
    "ANN", # flake8-annotations
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
]
ignore = [
    "ANN101",  # missing type annotation for self
    "ANN102",  # missing type annotation for cls
    "S104",    # possible binding to all interfaces
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101", "ANN"]  # Allow assert and missing annotations in tests

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--tb=short",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-report=xml:coverage.xml",
    "--cov-fail-under=80",
]

[tool.coverage.run]
source = ["app"]
omit = ["*/migrations/*", "*/tests/*", "app/main.py"]
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@(abc\\.)?abstractmethod",
]

[tool.bandit]
targets = ["app"]
skips = ["B101"]  # allow assert in non-test code where appropriate
severity = "medium"
confidence = "medium"
